requirements.txt
# Core dependencies
gradio==4.44.0
aiohttp==3.9.1
aiosqlite==0.19.0
aiofiles==23.2.1
python-dotenv==1.0.0

# LLM and AI
openai==1.3.7
huggingface-hub==0.19.4
transformers==4.36.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Utilities
python-multipart==0.0.6
requests==2.31.0
asyncio==3.4.3

# Optional: For better performance
uvloop==0.19.0; sys_platform != "win32"